fix hex

+++ ft_printf.h
@@ -6,7 +6,7 @@
 /*   By: tliangso <earth78203@gmail.com>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/09/14 00:26:58 by tliangso          #+#    #+#             */
-/*   Updated: 2022/09/14 22:31:40 by tliangso         ###   ########.fr       */
+/*   Updated: 2022/09/14 22:47:08 by tliangso         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -24,7 +24,7 @@ typedef struct s_sc
 size_t	ft_intlen(long int nb, int div);
 
 void	ft_printchr(va_list args, t_sc *sc);
-int		ft_printhex(size_t x, const char *format);
+int		ft_printhex(unsigned int x, const char *format);
 void	ft_printint(va_list args, t_sc *sc, const char *format);
 void	ft_printstr(va_list args, t_sc *sc);
 void	ft_hex(va_list args, t_sc *sc, const char *format);
